requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
sacrebleu>=2.3.1
bert-score>=0.3.13
nltk>=3.8.1
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
pyyaml>=6.0
pyarabic>=0.6.15
sentencepiece>=0.1.99
protobuf>=3.20.0
